Expose the download method
